+++ modules/specialist-directory-module/src/main/java/com/aidcompass/specialistdirectory/domain/review/models/dtos/ReviewResponseDto.java
@@ -14,7 +14,7 @@ public record ReviewResponseDto(
 
         String description,
 
-        long rating,
+        Integer rating,
 
         @JsonProperty("created_at")
         @JsonFormat(pattern = "yyyy-MM-dd HH:mm")
